backend-printing/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.2.0
astroid==3.1.0
async-timeout==4.0.3
attrs==23.2.0
az==0.1.0.dev1
azure-common==1.1.28
azure-core==1.29.4
azure-data-tables==12.5.0
azure-functions==1.3.0
azure-identity==1.15.0
azure-keyvault==4.2.0
azure-keyvault-certificates==4.7.0
azure-keyvault-keys==4.8.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.4.0
azure-storage-queue==12.9.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.4
Deprecated==1.2.14
dill==0.3.8
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.11.0
iniconfig==2.0.0
isodate==0.6.1
isort==5.13.2
marshmallow==3.21.0
marshmallow-dataclass==8.6.0
mccabe==0.7.0
microsoft-kiota-abstractions==1.2.0
microsoft-kiota-authentication-azure==1.0.0
microsoft-kiota-http==1.3.0
microsoft-kiota-serialization-json==1.0.1
microsoft-kiota-serialization-text==1.0.0
msal==1.26.0
msal-extensions==1.1.0
msgraph-core==1.0.0
msgraph-sdk==1.1.0
msrest==0.7.1
multidict==6.0.4
mypy-extensions==1.0.0
oauthlib==3.2.2
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
packaging==23.2
pendulum==3.0.0
platformdirs==4.2.0
pluggy==1.4.0
portalocker==2.8.2
pycparser==2.21
PyJWT==2.8.0
pylint==3.1.0
pytest==8.1.1
python-dateutil==2.8.2
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
std-uritemplate==0.0.52
time-machine==2.13.0
tomli==2.0.1
tomlkit==0.12.4
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2024.1
urllib3==2.0.5
wrapt==1.16.0
yarl==1.9.2
zipp==3.17.0